Extract types and decode into light weight libs (#10311)
