+++ core/services/relay/evm/mercury/utils/feeds_test.go
@@ -1,4 +1,4 @@
-package types
+package utils
 
 import (
 	"testing"
